% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSmort.R
\name{DBSmort}
\alias{DBSmort}
\title{Mortality Date Estimation}
\usage{
DBSmort(id, date, removed, zeta, eta, theta, filepath, keypath)
}
\arguments{
\item{id}{numeric. A specific collar id number. The user must select an individual.}

\item{date}{POSIXct object. The beginning date time in which to focus the search for defunct clusters. A date time must be listed in "YYYY-MM-DDTHH:MM:SS" format. Dates are downloaded in UTC+0.}

\item{removed}{POSIXct object. The last date time in which to consider when searching for defunct clusters. The date should be moved forward if collars were moved from the study site and a secondary cluster was created. A date time must be listed in "YYYY-MM-DDTHH:MM:SS" format. Dates are downloaded in UTC+0.}

\item{zeta}{numeric. Threshold (determined by the CRS, typically meters) for determining non-movement and the formation of a defunct cluster. For example, if 10 is specified, fixes less than or equal to 10 meters would be considered non-movement. See ?flag_defunct_clusters for more information.}

\item{eta}{numeric. The minimum number of steps to be considered non-movement. For example, 2 would mean the animal must persist in a state of non-movement for a minimum of two success fixes to create a defunct cluster. See ?flag_defunct_clusters for more information.}

\item{theta}{numeric. The minimum number of hours in which a cluster must persist in a non-movement state. For example, 8 would mean the animal must meet the aforementioned criteria for a minimum of 8 hours to create a defunct cluster. See ?flag_defunct_clusters for more information.}

\item{filepath}{File path to the locally stored metadata. The user must specify a file path.}

\item{keypath}{File path to the locally stored Vectronic collar keys. Folder should only contain keys of deployed collars. The user must specify a file path.}
}
\description{
This package and the functions within were developed to assist collaborators on the Sonora Mexico desert bighorn sheep project identify defunct clusters and accurately predict mortality and/or dropped collar dates. The package relies on functions in the amt package that can be adjusted to better identify defunct clusters. The user can also specify if sections of location data should be ignored. For example, collars are often recovered and powered off hours to weeks post-recovery. This can create secondary clusters that traditional tools are unable to account for. This package allows the user to filter out these clusters to better identify true defunct clusters. The user can select the option to have the updated metadata file output in CSV format. A new folder will be created in the directory that contains the original metadata. Additionally, the user will select prompts within the code to better tailor the tool to the specific project. Lastly, an interactive map is displayed with live fixes, defunct cluster fixes, movement path, and a predicted final live fix to allow the user to confirm the validity of the predicted dates before data is extracted. The function is specifically designed to pull location data from the API using Vectronic collar keys. As such, all time stamps are in UTC+0.
}
\examples{
DBSmort(id = 39971,
         date = '2023-02-10T00:00:01',
         removed = '2023-02-15T21:01:39',
         zeta = 30,
         eta = 2,
         theta = 6,
         filepath = "E:\\\\Texas A&M\\\\DBS Mexico\\\\metadata files\\\\metadata.xlsx",
         keypath = "E:\\\\Texas A&M\\\\DBS Mexico\\\\Vectronic Key Files")  #' theta is in hours!
}
\author{
Dylan Stewart, Ph.D. student, Texas A&M University, dylan.stewart@tamu.edu
}
