% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSvideo.R
\name{DBSvideo}
\alias{DBSvideo}
\title{Collar Video File Reorganization}
\usage{
DBSvideo(
  video,
  mortdata,
  id,
  selection,
  begin,
  end,
  videofolderpath,
  veckeypath,
  metadatafilepath,
  savevideometadata,
  subdivide
)
}
\arguments{
\item{video}{logical. If TRUE, then the paired video locations and GPS fixes will both be projected on an interactive map. If FALSE, only the GPS data will be projected. The user must specify TRUE or FALSE.}

\item{mortdata}{logical. If TRUE, mortality data will be displayed and color coded on an interactive map. If FALSE, mortality data will not be color coded. Only specify TRUE if mortality data is available for the individual. The user must specify TRUE or FALSE.}

\item{id}{numeric. GPS data will be mapped for the individual. Only one individual may be selected. The user must specify an individual.}

\item{selection}{logical. If TRUE, mapping will range from the date time specified between "begin" and "end". If FALSE, all GPS data will be included in the interactive map. The user must specify TRUE or FALSE.}

\item{begin}{POSIXct object. If selection is TRUE, mapping will start at the date time specified. Date time objects must be in "YYYY-MM-DD HH:MM:SS" format. The time stamp object must contain parentheses (e.g.,"2023-01-01 00:00:01"), and all date time objects are subset based on UTC+0. The user must specify a date time if selection == TRUE.}

\item{end}{POSIXct object. If selection is TRUE, mapping will end at the date time specified. Date time objects must be in "YYYY-MM-DD HH:MM:SS" format. The time stamp object must contain parentheses (e.g.,"2023-01-01 00:00:01"), and all date time objects are subset based on UTC+0. The user must specify a date time if selection == TRUE.}

\item{videofolderpath}{File path to the locally stored collar video file containing folders for each individual. Video files for each individual must be in id_collarid format. The user must specify a file path.}

\item{veckeypath}{File path to the locally stored Vectronic collar keys. Folder should only contain keys of deployed collars. The user must specify a file path.}

\item{metadatafilepath}{File path to the locally stored metadata. User should verify that the metadata is up-to-date. The user must specify a file path.}

\item{savevideometadata}{logical. If selection is TRUE, a copy of the video metadata will be exported to the video folder. If FALSE, the metadata will not be exported. The user must specify TRUE or FALSE.}

\item{subdivide}{logical. If selection is TRUE, video files will be copied into new folders based on id, video status, and coordinate acquisition success. The user must specify TRUE or FALSE.}
}
\description{
This package and the functions within were developed to assist collaborators on the Sonora Mexico desert bighorn sheep project reorganize video files (.mp4) collected from GPS collars into new subdirectories based on fix success and coordinate acquisition. Multiple options were added to allow the user to visualize the GPS data and paired video acquisition locations interactively. These options allow the user to specify the inclusion or exclusion of mortality data and/or video location data. Interactive mapping is specific to an individual, which the user can specify within the function. The collaborator also has the option to select between a range of dates or view all the GPS data for the individual. All dates are in UTC+0. Lastly, the user can specify to have a copy of the video metadata exported and the video files copied and reorganized in folders based on fix success and coordinate acquisition. It is imperative that the user select a choice for each option in the function.
}
\examples{
DBSvideo(video = TRUE,
         mortdata = TRUE,
         id = 39971,
         selection = FALSE,
         begin = "2023-02-01 00:00:01",
         end = "2023-02-20 00:00:01",
         videofolderpath = "E:\\\\Texas A&M\\\\DBS Mexico\\\\Video",
         veckeypath = "E:\\\\Texas A&M\\\\DBS Mexico\\\\Vectronic Key Files",
         metadatafilepath = "E:\\\\Texas A&M\\\\DBS Mexico\\\\metadata files\\\\metadata.xlsx",
         savevideometadata = TRUE,
         subdivide = TRUE)
}
\author{
Dylan Stewart, Ph.D. student, Texas A&M University, dylan.stewart@tamu.edu
}
